+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added support PDF/A and PDF/UA
+- Changed Virtual file system (VFS) format for better compatibility with frameworks (backwards compatibility preserved). **For compatibility with frameworks, rebuild VFS required!**
 - Update pdfkit to 0.15.1
 - Fixed bug with how page breaks provoked by cells with rowspan were handled
 - Fixed find where previous cell started with row span and col span combination